Import to the new hooks API
In WordPoints 2.1.0 a new hooks API is being introduced, so we should
import CubePoints’s points settings to that instead of to the old
points hooks. In order to do this, it is necessary for us to make a few
changes in the way that we import the points logs as well. First, we
need to make the imported log types unique to the type of transaction,
instead of just giving the all the type `cubepoints`. We also need to
properly set the `auto_reversed` and `original_log_id` metadata for
reversible transactions. And we need to save the ID of the primary
entity each imported log relates to as metadata as well.

In the process, we fix a bug where we were importing points hooks for
all points types for the post publish event, even though CubePoints
only awarded points for the post points type.

We also include tests that run both CubePoints and the imported hook
against the same code, to ensure that capable behavior is maintained
with the imported hooks.

Fixes #8