Respect existing periods in imported reactions
Utilize the new points legacy period extension in order to continue
respected started periods even after the hooks are imported.

We also import the periods to be relative instead of absolute, since
they are relative in CubePoints.

In order for us to test this properly, it is necessary for us to
activate the daily points module when we load CubePoints in our
bootstrap. This causes it to award points during the bootstrap loading,
which means we have to tweak the tests in a few places.

We also introduce similar loading of the post author module, and update
our tests relating to it accordingly.

See #8