Remove unneeded workaround
We thought that we needed this, but it was actually due to a local
issue with some cruft left over in our test DB. Our Travis build was
failing, so we tried dropping all the tables locally and starting from
scratch, and we found that these two lines were causing the two failing
tests in Travis. Removing them and all is well!

See #8