Fix PHPUnit tests on WP 4.7
When running against WordPoints 2.1 on WordPress 4.7 the tests were ending in a fatal error from one of the functions from the modules not existing.. This was due to the modules not being loaded, because in WordPress 4.7 hooks that are hooked to an earlier priority of the same hook that is currently running do not run, whereas in earlier versions those functions hooked up late would be run at the end. The function that loads the CubePoints modules is hooked to `plugins_loaded` with a priority of 2, however, we do not load CubePoints at all until priority 14, and so that function was being hooked up too late. We now just call it manually on newer versions of WordPress. This special handling is only needed for WordPoints 2.1 anyway, and will be removed in later versions, where we can just use WPPPB instead.